changed components on project form Set to List (for sorting) and added some service methods/DTO-fields

+++ web-client/trunk/src/main/java/eu/cityopt/DTO/MetricValDTO.java
@@ -5,7 +5,7 @@ import lombok.Setter;
 
 public class MetricValDTO {
 
-	@Getter @Setter private long metricvalid;
+	@Getter @Setter private int metricvalid;
 //	@Getter @Setter private Metric metric;
 //	@Getter @Setter private ScenarioMetrics scenariometrics;
 //	@Getter @Setter private TimeSeries timeseries;
